Machine learning/requirement.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
tensorflow
opencv-python
nltk
spacy
textblob
emoji
cleantext
gensim
scipy==1.12
face-detection
ipython
transformers
huggingface_hub
gtts
langchain_community
pypdf
bs4
google-generativeai
langchain_google_genai
chromadb
langchain_huggingface
sentence-transformers
faiss-cpu
langcain-chroma